requirements.txt
streamlit==1.32.0
faster-whisper==0.10.0
google-generativeai==0.3.2
moviepy==1.0.3
torch==2.2.0
python-dotenv==1.0.0
streamlit-extras==0.4.0
streamlit-option-menu==0.3.12
watchdog==4.0.0 